+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/payments/InvoiceContent.kt
@@ -1,17 +1,14 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.message.payments
 
 import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.Request
-import com.github.insanusmokrassar.TelegramBotAPI.requests.send.payments.SendInvoice
 import com.github.insanusmokrassar.TelegramBotAPI.types.ChatIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.MessageIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.buttons.KeyboardMarkup
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.ContentMessage
-import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.Message
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.content.abstracts.MessageContent
-import com.github.insanusmokrassar.TelegramBotAPI.types.message.payments.abstracts.PaymentInfo
 import com.github.insanusmokrassar.TelegramBotAPI.types.payments.Invoice
 
-data class InvoiceOfPayment(
+data class InvoiceContent(
     val invoice: Invoice
 ) : MessageContent {
     override fun createResend(
@@ -19,7 +16,7 @@ data class InvoiceOfPayment(
         disableNotification: Boolean,
         replyToMessageId: MessageIdentifier?,
         replyMarkup: KeyboardMarkup?
-    ): Request<ContentMessage<InvoiceOfPayment>> {
+    ): Request<ContentMessage<InvoiceContent>> {
         error("Unfortunately, currently InvoiceOfPayment can not be resend due to requirement of additional parameters," +
             " which can't be provided during the call of this method")
     }